demo/py.examples/requirements-32.txt
beautifulsoup4
openpyxl
Requests
pillow
plotly
pandas